UGC-57 Use endpoint to retrieve and show all modules in the form (frontend)
- Fetch modules from API and group them by level (keys '4', '5', '6')
- Map API properties: moduleCode -> code and moduleName -> name
- Added error logging for module data fetch failures
