+++ frontend/index.html
@@ -11,6 +11,7 @@
   <link rel="stylesheet" href="css/styles.css">
 
   <!-- JS files in order -->
+  <script src="https://cdn.jsdelivr.net/npm/axios/dist/axios.min.js"></script>
   <script src="js/data/modulesByLevel.js" defer></script>
   <script src="js/services/StorageService.js" defer></script>
   <script src="js/services/MessageService.js" defer></script>
