+++ frontend/js/app.js
@@ -5,31 +5,6 @@
  * The DOMContentLoaded event ensures that the script runs after the document is fully loaded.
  * This is crucial for ensuring that all elements are available for manipulation.
  */
-
-import axios from 'axios';
-
-// Function to fetch modules and organize them by level
-const fetchModules = async () => {
-  try {
-    // Make GET request to backend endpoint
-    const response = await axios.get('http://localhost:3000/api/modules');
-    const modules = response.data;
-
-    // Organize modules into modulesByLevel variable
-    const modulesByLevel = {};
-    Object.keys(modules).forEach(level => {
-      modulesByLevel[level] = modules[level];
-    });
-
-    console.log('Modules organized by level:', modulesByLevel);
-  } catch (error) {
-    console.error('Error fetching modules:', error);
-  }
-};
-
-// Call the function to test connection and fetch data
-fetchModules();
-
 document.addEventListener("DOMContentLoaded", () => {
     const storageService = new StorageService();
     const messageDisplay = document.getElementById("messageDisplay");
