+++ frontend/js/data/modulesByLevel.js
@@ -1,43 +1,70 @@
 /**
- * * This file contains the module data for each level of the Computer Science course.
- * * Each module is represented by an object with a code and name.
- * * This data is used to populate the module selection form in the application.
- * * The data is structured as an object where the keys are level numbers (as strings),
+ * @file modulesByLevel.js
+ * @description Manages Computer Science module data organized by academic level (4, 5, 6)
+ * 
+ * Provides:
+ * - API fetching of module data
+ * - Organization by level with string keys ("4", "5", "6")
+ * - Consistent property naming (code/name)
+ * - Error handling and logging
+ * 
+ * Expected API Data Format:
+ * [
+ *   {
+ *     level: number,              // Academic level (4, 5, or 6)
+ *     moduleCode: string,         // Official module code (e.g. "CM1005")
+ *     moduleName: string          // Full module title
+ *   },
+ *   ...
+ * ]
+ * 
+ * Output Structure:
+ * {
+ *   "4": [{code: string, name: string}, ...],  // Level 4 modules
+ *   "5": [{code: string, name: string}, ...],  // Level 5 modules  
+ *   "6": [{code: string, name: string}, ...]   // Level 6 modules
+ * }
+ * 
+ * @requires axios
+ * @exports modulesByLevel
  */
-const modulesByLevel = {
-  "4": [
-      { code: "CM1005", name: "Introduction to Programming 1" },
-      { code: "CM1010", name: "Introduction to Programming 2" },
-      { code: "CM1015", name: "Computational Mathematics" },
-      { code: "CM1020", name: "Discrete Mathematics" },
-      { code: "CM1030", name: "How Computers Work" },
-      { code: "CM1035", name: "Algorithms and Data Structures" },
-      { code: "CM1040", name: "Web Development" },
-      { code: "CM1025", name: "Fundamentals of Computer Science" }
-  ],
-  "5": [
-      { code: "CM2045", name: "Professional Practice for Computer Science" },
-      { code: "CM2035", name: "Algorithms and Data Structures 2" },
-      { code: "CM2025", name: "Computer Security" },
-      { code: "CM2040", name: "Databases, Networks and the Web" },
-      { code: "CM2030", name: "Graphics Programming" },
-      { code: "CM2005", name: "Object Oriented Programming" },
-      { code: "CM2015", name: "Programming with Data" },
-      { code: "CM2010", name: "Software Design and Development" }
-  ],
-  "6": [
-      { code: "CM3045", name: "3D Graphics and Animation" },
-      { code: "CM3035", name: "Advanced Web Development" },
-      { code: "CM3020", name: "Artificial Intelligence" },
-      { code: "CM3005", name: "Data Science" },
-      { code: "CM3010", name: "Databases and Advanced Data Techniques" },
-      { code: "CM3030", name: "Games Development" },
-      { code: "CM3065", name: "Intelligent Signal Processing" },
-      { code: "CM3055", name: "Interaction Design" },
-      { code: "CM3015", name: "Machine Learning and Neural Networks" },
-      { code: "CM3050", name: "Mobile Development" },
-      { code: "CM3060", name: "Natural Language Processing" },
-      { code: "CM3040", name: "Physical Computing and Internet of Things" },
-      { code: "CM3025", name: "Virtual Reality" }
-  ]
+
+// Global reference to organized modules
+const modulesByLevel = {};
+
+/**
+ * Fetches and organizes modules from API
+ * @async
+ * @function fetchModules
+ * @description Retrieves module data from API and structures it by level
+ * - Maintains old data structure using new data from API
+ * - Preserves original property naming (code/name)
+ * - Groups modules under string-based level keys
+ * - Handles and logs errors without breaking execution
+ */
+const fetchModules = async () => {
+  try {
+    const response = await axios.get('http://localhost:3000/api/modules');
+    const modules = response.data;
+
+    // Process each module: group by level and transform properties
+    modules.forEach(module => {
+      const levelKey = module.level.toString(); // Ensure string key format
+      if (!modulesByLevel[levelKey]) {
+        modulesByLevel[levelKey] = []; // Initialize array for new levels
+      }
+      modulesByLevel[levelKey].push({
+        code: module.moduleCode,  // Map moduleCode → code
+        name: module.moduleName   // Map moduleName → name
+      });
+    });
+
+    console.log('Module data successfully loaded');
+
+  } catch (error) {
+    console.error('Module data fetch failed:', error);
+  }
 };
+
+// Initial data load
+fetchModules();
